[tensor] Add scalar in tensor
What I did:
- add scalar whose dim_size is 0, shape is 1
- add helper methods: is_scalar, scalar_value,
  set_scalar_value
- add help function: make_scalar

+++ src/kernel/cpu/softmax_kernel_cpu.cpp
@@ -39,7 +39,7 @@ void softmax_kernel_cpu(const tensor::Tensor& input, void* stream) {
  *
  * @note The function modifies the input array in-place
  */
-void softmax_kernel_cpu(const float* input_ptr, size_t size) {
+void softmax_kernel_cpu(const float* input_ptr, int32_t size) {
   tensor::Tensor input(core::DataType::FP32, size);
   std::shared_ptr<core::Buffer> buffer =
       std::make_shared<core::Buffer>(size * sizeof(float), nullptr, (void*)input_ptr, true);
